zynqmp_pmufw: Fix violations for MISRA Rule 10_3
This patch fix following violation
misra_c_2012_rule_10_3_violation: The value of an expression shall
not be assigned to an object with a narrower essential type or
of a different essential type category.

To solve this MISRA violation do explicit type cast as per requirement.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_gic_proxy.c
@@ -116,7 +116,7 @@ static void PmGicProxyDisable(void)
 	/* Disable FPD GPI1 wake event */
 	DISABLE_WAKE(PMU_LOCAL_GPI1_ENABLE_FPD_WAKE_GIC_PROX_MASK);
 
-	pmGicProxy.flags &= ~PM_GIC_PROXY_IS_ENABLED;
+	pmGicProxy.flags &= ~(u8)PM_GIC_PROXY_IS_ENABLED;
 }
 
 /**
